refactor components naming and rendering startegies to more react way
