+++ src/lib/If/index.tsx
@@ -1,9 +1,9 @@
-interface IProps {
+interface TypeProps {
   children: React.ReactNode;
   is: boolean;
 }
 
-function If({ is, children }: IProps): JSX.Element {
+function If({ is, children }: TypeProps): JSX.Element {
   return <>{is ? children : null}</>;
 }
