--- src/lib/components/For/index.tsx
@@ -1,12 +1,0 @@
-interface IProps<T> {
-  children: (el: T, idx: number) => React.ReactNode;
-  each: T[];
-}
-
-function For<T>({ children, each }: IProps<T>) {
-  const rendered = each.map((el, idx) => children(el, idx));
-
-  return rendered;
-}
-
-export default For;
