--- src/lib/components/Loop/index.tsx
@@ -1,16 +1,0 @@
-interface IProps {
-  children: React.ReactNode;
-  times: number;
-}
-
-function Loop({ times, children }: IProps): JSX.Element {
-  const rendered = [];
-
-  for (let i = 0; i < times; i++) {
-    rendered.push(children);
-  }
-
-  return <>{rendered}</>;
-}
-
-export default Loop;
